+++ main/ABPMRR.h
@@ -3,8 +3,8 @@
 
 #include <inttypes.h>
 #include <stdio.h>
-// #include "I2C.h"
 #include <string.h>
+#include <logdef.h>
 #include "Setup.h"
 #include <math.h>
 #include "I2Cbus.hpp"
@@ -44,7 +44,7 @@ class ABPMRR : public AirspeedSensor
     // public functions
         bool  doOffset( bool force=false );
         float readPascal( float minimum, bool &ok );
-        bool  offsetPlausible( uint16_t offset );
+        bool  offsetPlausible( uint32_t offset );
         bool  selfTest( int& adval );
         void  setBus( I2C_t *theBus ) {  bus = theBus; };
         void  changeConfig();
